Re-organize the code preparing for other backends
