--- MANIFEST.in
@@ -1,7 +1,0 @@
-include pyproject.toml
-include LICENSE
-
-recursive-include src *
-recursive-include include *
-include CMakeLists.txt
-include VERSION
