+++ python/vesin/MANIFEST.in
@@ -1,0 +1,5 @@
+include pyproject.toml
+include LICENSE
+include VERSION
+
+recursive-include lib *
