+++ python/vesin/VERSION
@@ -1,0 +1,1 @@
+../../vesin/VERSION
