+++ python/vesin/vesin/_neighbors.py
