+++ tox.ini
@@ -20,7 +20,7 @@ deps =
     wheel
 
 commands =
-    pip wheel . --no-deps --no-build-isolation --check-build-dependencies --wheel-dir {envtmpdir}/dist
+    pip wheel python/vesin --no-deps --no-build-isolation --check-build-dependencies --wheel-dir {envtmpdir}/dist
 
 [testenv:tests]
 description = Run the tests of the Python package
@@ -28,6 +28,7 @@ deps =
     ase
     pytest
 
+changedir = python/vesin
 commands =
     pytest {posargs}
 
@@ -39,7 +40,7 @@ passenv = *
 deps = cmake
 
 commands =
-    cmake -B {envtmpdir} -S . -DVESIN_BUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Debug
+    cmake -B {envtmpdir} -S vesin -DVESIN_BUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Debug
     cmake --build {envtmpdir} --config Debug
     ctest --test-dir {envtmpdir} --build-config Debug
