+++ vesin/src/math.hpp
